refactor: Modularize playground code
Modularized the 863-line +page.svelte into separate modules:

New modules:
- $lib/stores/playground.svelte.js (592 lines) - Centralized state management using Svelte 5 runes
- $lib/utils/ast.js (84 lines) - AST normalization utilities
- $lib/utils/share.js (42 lines) - Share link utilities
- $lib/utils/workspace.js (97 lines) - Workspace tree utilities
- $lib/runtime/runner.svelte.js (69 lines) - Reactive worker runner

Updated:
- $lib/editor/mimo-theme.js - Now uses CSS variables for dynamic theming, matches playground UI
- +page.svelte - Reduced from 863 to 173 lines (80% reduction)

Benefits:
- Separation of concerns (state, utilities, UI)
- Testable utility functions
- Reusable modules
- Dynamic editor theming via CSS variables
- Cleaner, more maintainable code structure
